requirements.txt
gradio>=4.0.0
openai-whisper
ffmpeg-python
torch